Correct code for initializing msalInstance
Text links to https://github.com/AzureAD/microsoft-authentication-library-for-js/blob/dev/lib/msal-browser/docs/initialization.md which says you have to initialize the instance, but this sample code doesn't initialize. Either needs to initialize it, or if it doesn't then we need to update the text to explain e.g. that MsalProvider initializes it automatically.